+++ dev_iot_core/core/Timer.h
@@ -3,14 +3,14 @@
 
 class Timer {
 private:
-  uint16_t timeout = 3000;
-  uint16_t prevMillis = 0;
+  uint32_t timeout = 3000;
+  uint32_t prevMillis = 0;
   bool _lock = false;
 
 public:
   Timer(){};
 
-  Timer(uint16_t timeout)
+  Timer(uint32_t timeout)
     : timeout(timeout){};
 
   bool isExpired(void) {
@@ -21,7 +21,7 @@ public:
     this->prevMillis = millis();
   }
 
-  bool autoTimeout(uint16_t _to) {
+  bool autoTimeout(uint32_t _to) {
     bool isTimeout = millis() - this->prevMillis > _to;
 
     if (this->_lock == false) {
